organizing the content
