+++ src/components/AccordionRecyclable/accordionRecyclable.scss
