+++ src/components/Accordions/Accordions.scss
@@ -1,0 +1,7 @@
+.body-accordions {
+    align-items: center;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    margin: 60px 0;
+  }
