+++ src/components/DownloadPdf/index.js
@@ -1,0 +1,14 @@
+import { IconLink } from '@tabler/icons';
+
+const DownloadPdf = () => {
+    return(
+        <p>Para mayor información…
+                  <a href='/' download="./assets/pdf/Optihelp-mayor-informacion.pptx" >
+                    <IconLink />
+                    PPTX
+                  </a>
+                </p>
+    );
+}
+
+export { DownloadPdf }
